+++ tasks/connect.js
@@ -1,0 +1,8 @@
+var gulp = require("gulp");
+var connect = require("gulp-connect");
+
+gulp.task("connect", function () {
+  connect.server({
+    root: "build"
+  });
+});
