+++ tasks/default.js
@@ -1,0 +1,3 @@
+var gulp = require('gulp');
+
+gulp.task("default", ["style", "news", "pages", "work", "connect", "watch"]);
